Move contrib attention operators to a separate file
Move the `contrib` submodule of `src/ops/attention.rs` to
`src/ops/attention/contrib.rs` and move the tests for the contrib
operators into a `tests` module in that file. This reduces the size of
attention.rs from ~3.8K lines to ~2K.
